Create to_do_list_with_gui
This is a To-Do-List Application
This is built with Tinkter a GUI Library
This allowed you to add tasks, remove tasks and see your tasks.
The tasks will save automatically when the program is closed to a text.txt file. 
When the program is loaded again the saved tasks will appear.